backend/requirements.txt
pymongo[srv]==4.10.1
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
pydantic-settings
pytest
httpx
